컴파일 경고 수정

+++ ICAA/ThreadTask/DFTaskMngr.cpp
@@ -2145,7 +2145,7 @@ int CDFTaskMngr::GetCloseDFFreq(UINT Freq)
 			break;
 		}
 
-		if( nFreq >=  Freq)  
+		if( nFreq >= (int)Freq )  
 		{
 			bigDiff = abs(nFreq  - (int)Freq);
 			break;
